--- source/pyspark_functions/release-notes/release-notes.md
@@ -1,4 +1,0 @@
-# Version 0.0.1
-
-- Bootstraped project.
-- Added functions and tests.
